TiepTinTuc

+++ app/Http/Controllers/TinTucController.php
@@ -4,17 +4,32 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\TinTuc;
+use Illuminate\Support\Facades\Gate;
 
 class TinTucController extends Controller
 {
+    public function HienThi()
+    {
+        $news = TinTuc::all(); 
+
+        return view('user/tin-tuc/hien-thi', compact('news'));
+    }
+    public function ChiTietHienThi($id)
+    {
+        $news = TinTuc::findOrFail($id);
+        return view('user/tin-tuc/chi-tiet', compact('news'));
+    }
     public function DanhSach()
     {
-        $news = TinTuc::latest()->paginate(10); // Lấy danh sách tin tức, sắp xếp theo thời gian và phân trang
+        $news = TinTuc::latest()->paginate(10);
         return view('admin/tin-tuc/danh-sach', compact('news'));
     }
 
     public function ThemMoi()
     {
+        if (Gate::denies('quan-ly-tin-tuc')) {
+            return redirect()->route('admin.index')->with('error','bạn không có quyền truy cập vào chức năng này');
+        }
         return view('admin/tin-tuc/them-moi');
     }
 
@@ -22,13 +37,16 @@ class TinTucController extends Controller
     {
         $request->validate([
             'tieu_de' => 'required',
+            'mo_ta_ngan' => 'required',
             'noi_dung' => 'required',
             'image' => 'image|mimes:jpeg,png,jpg,gif|max:2048', // Kiểm tra hình ảnh có đúng định dạng và kích thước
         ]);
 
         $news = new TinTuc();
         $news->tieu_de = $request->tieu_de;
+        $news->mo_ta_ngan=$request->mo_ta_ngan;
         $news->noi_dung = $request->noi_dung;
+        $news->ngay_dang = now();
 
         // Xử lý lưu hình ảnh nếu có được tải lên
         if ($request->hasFile('image')) {
@@ -45,27 +63,36 @@ class TinTucController extends Controller
     }
     public function ChiTiet($id)
     {
+        if (Gate::denies('quan-ly-tin-tuc')) {
+            return redirect()->route('admin.index')->with('error','bạn không có quyền truy cập vào chức năng này');
+        }
         $news = TinTuc::findOrFail($id);
-        return view('tin-tuc.chi-tiet', compact('news'));
+        return view('admin/tin-tuc/chi-tiet', compact('news'));
     }
 
     public function CapNhat($id)
     {
+        if (Gate::denies('quan-ly-tin-tuc')) {
+            return redirect()->route('admin.index')->with('error','bạn không có quyền truy cập vào chức năng này');
+        }
         $news = TinTuc::findOrFail($id);
-        return view('tin-tuc.cap-nhat', compact('news'));
+        return view('admin/tin-tuc/cap-nhat', compact('news'));
     }
 
     public function XuLyCapNhat(Request $request, $id)
     {
         $request->validate([
             'tieu_de' => 'required',
+            'mo_ta_ngan' => 'required',
             'noi_dung' => 'required',
             'image' => 'image|mimes:jpeg,png,jpg,gif|max:2048', // Kiểm tra hình ảnh có đúng định dạng và kích thước
         ]);
 
         $news = TinTuc::findOrFail($id);
         $news->tieu_de = $request->tieu_de;
+        $news->mo_ta_ngan=$request->mo_ta_ngan;
         $news->noi_dung = $request->noi_dung;
+        $news->ngay_dang = now();
 
         // Xử lý lưu hình ảnh nếu có được tải lên
         if ($request->hasFile('image')) {
@@ -83,6 +110,9 @@ class TinTucController extends Controller
 
     public function Xoa($id)
     {
+        if (Gate::denies('quan-ly-tin-tuc')) {
+            return redirect()->route('admin.index')->with('error','bạn không có quyền truy cập vào chức năng này');
+        }
         $news = TinTuc::findOrFail($id);
         $news->delete();
